Update Auth to Access button and mobile styles
This change adds a GitHub logo to the "Sign in With GitHub" button rendered with
the trails, as well as fixing some mobile styling issues for the step preview
descriptions on mobile.

Also, as part of this change, the `videos/_video_for_trail_preview` and
`exercises/_exercises_for_trail_preview` partials have been merged into
`trails/_completeable_preview` which removes a good amount of duplication and
prevents potential slips when adding functionality (link / no link based on user
access).
