+++ app/assets/stylesheets/_exercises-and-teachers.scss
@@ -20,13 +20,20 @@
       width: 100%;
 
       .exercise-title {
+        @include dashboard-small-only {
+          margin-left: 50px;
+        }
+
         margin-bottom: 0.5em;
       }
     }
 
     .exercise-item {
       display: inline-block;
-      margin-left: 80px;
+
+      @include dashboard-medium {
+        margin-left: 50px;
+      }
     }
   }
 
@@ -44,9 +51,16 @@
   }
 
   .available-link {
-    border: 1px solid $upcase-green;
     border-radius: $base-border-radius;
+    border: 1px solid $upcase-green;
     color: $upcase-green;
+    display: inline-block;
     padding: 5px 10px;
+
+    .logo {
+      max-height: 1.4em;
+      text-align: center;
+      vertical-align: bottom;
+    }
   }
 }
