--- app/views/exercises/_exercise_for_trail_preview.html.erb
@@ -1,18 +1,0 @@
-<figure class="exercise <%= exercise.state.parameterize %>-exercise">
-  <div class="dot-container">
-    <div class="dot"></div>
-  </div>
-
-  <% if current_user_has_active_subscription? %>
-    <%= link_to exercise.url, class:"exercise-item", title: exercise.name do %>
-      <h3><%= exercise.name %></h3>
-      <p><%= exercise.summary %></p>
-    <% end %>
-  <% else %>
-    <div class="exercise-item">
-      <h3><%= exercise.name %></h3>
-      <p><%= exercise.summary %></p>
-      <%= render "products/locked", feature: exercise %>
-    </div>
-  <% end %>
-</figure>
