--- app/views/videos/_video_for_trail_preview.html.erb
@@ -1,18 +1,0 @@
-<figure class="exercise <%= video.state.parameterize %>-exercise">
-  <div class="dot-container">
-    <div class="dot"></div>
-  </div>
-
-  <% if current_user_has_access_to?(video) %>
-    <%= link_to video_path(video), class:"exercise-item", title: video.name do %>
-      <h3><%= video.name %></h3>
-      <p><%= video.summary %></p>
-    <% end %>
-  <% else %>
-    <div class="exercise-item">
-      <h3><%= video.name %></h3>
-      <p><%= video.summary %></p>
-      <%= render "products/locked", feature: video %>
-    </div>
-  <% end %>
-</figure>
