+++ spec/support/trail_helpers.rb
@@ -1,0 +1,5 @@
+module TrailHelpers
+  def build_completeable_with_progress(state: Status::UNSTARTED)
+    CompleteableWithProgress.new(build_stubbed(:video), state)
+  end
+end
